Update speaker-example to work with wrapper, renamed test_wrapper
