+++ examples/output-rows.js
@@ -1,7 +1,6 @@
 OpenMTP_Module = require('../index.js');
 
 const fs = require('fs');
-const Speaker = require('speaker');
 
 fs.readFile('alf2_zalza_edit.xm', function(err, data) {
     if(err != null) {
@@ -10,30 +9,6 @@ fs.readFile('alf2_zalza_edit.xm', function(err, data) {
         var module = new OpenMTP_Module(data);
         
         console.log(module)
-
-        console.log('Position', module.position_seconds);
-        console.log('Repeat', module.repeat);
-        console.log('Duration', module.duration_seconds);
-        console.log('Instruments', module.num_instruments);
-        console.log('Tempo', module.current_tempo);
-        console.log('Speed', module.current_speed);
-        console.log('Input Channels', module.num_channels);
-
-        console.log(module.metadata);
-        
-        /*
-        var speakerStream = new Speaker();
-        var chiptuneStream = module.openAsStream();
-        chiptuneStream.pipe(speakerStream);
-        
-        
-        chiptuneStream.on('close', function(){
-			chiptuneStream.unpipe();
-			chiptuneStream.destroy();
-			module.destroy();
-			module = null;
-		})
-		*/
 		
 		for(var order = module.current_order; order < module.num_orders; order++) {
 			const pattern = module.get_order_pattern(order);
